requirements.txt
pillow==11.0.0
qrcode==8.0
pandas==2.2.3
zebra==0.1.0
zebrafy==1.2.2
openpyxl==3.1.5
python-dotenv==1.0.1